fixed duplicated type issue
