+++ src/results/resultCell.ts
@@ -1,0 +1,36 @@
+/**
+ * Copyright 2021 RelationalAI, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy
+ * of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+
+import { ResultColumnDef } from './resultColumn';
+import { toDisplayValue, toJsValue } from './resultUtils';
+
+export type CellToJsFormat = 'cell' | 'value' | 'displayValue' | 'rawValue';
+
+export class ResultCell {
+  constructor(private arrowValue: any, public columnDef: ResultColumnDef) {}
+
+  get rawValue() {
+    return this.arrowValue;
+  }
+
+  get value() {
+    return toJsValue(this.arrowValue, this.columnDef.type);
+  }
+
+  get displayValue() {
+    return toDisplayValue(this.value, this.columnDef.type);
+  }
+}
